Refactor SCA transformation by target to have clearer code
