+++ cx2t/src/claw/wani/transformation/sca/Parallelize.java
@@ -9,9 +9,7 @@ import claw.shenron.translator.Translator;
 import claw.tatsu.common.*;
 import claw.tatsu.directive.common.Directive;
 import claw.tatsu.primitive.*;
-import claw.tatsu.xcodeml.abstraction.AssignStatement;
 import claw.tatsu.xcodeml.abstraction.DimensionDefinition;
-import claw.tatsu.xcodeml.abstraction.NestedDoStatement;
 import claw.tatsu.xcodeml.abstraction.PromotionInfo;
 import claw.tatsu.xcodeml.exception.IllegalTransformationException;
 import claw.tatsu.xcodeml.xnode.XnodeUtil;
@@ -20,11 +18,8 @@ import claw.tatsu.xcodeml.xnode.fortran.*;
 import claw.wani.language.ClawPragma;
 import claw.wani.transformation.ClawTransformation;
 import claw.wani.x2t.configuration.Configuration;
-import claw.wani.x2t.configuration.AcceleratorConfiguration;
-import claw.wani.x2t.configuration.AcceleratorLocalStrategy;
 
 import java.util.*;
-import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * The Single Column Abstraction (SCA) transformation transforms the code
@@ -79,13 +74,13 @@ import java.util.concurrent.atomic.AtomicInteger;
  */
 public class Parallelize extends ClawTransformation {
 
-  private final Map<String, DimensionDefinition> _dimensions;
-  private final Map<String, PromotionInfo> _promotions;
-  private final Set<String> _arrayFieldsInOut;
-  private final Set<String> _scalarFields;
-  private int _overDimensions;
-  private FfunctionDefinition _fctDef;
-  private FfunctionType _fctType;
+  protected final Map<String, DimensionDefinition> _dimensions;
+  protected final Map<String, PromotionInfo> _promotions;
+  protected final Set<String> _arrayFieldsInOut;
+  protected final Set<String> _scalarFields;
+  protected int _overDimensions;
+  protected FfunctionDefinition _fctDef;
+  protected FfunctionType _fctType;
 
   /**
    * Constructs a new Parallelize transformation triggered from a specific
@@ -360,26 +355,18 @@ public class Parallelize extends ClawTransformation {
     }
 
     removePragma();
+  }
 
-    // Apply specific target transformation
-    if(Context.isTarget(Target.GPU)) {
-      transformForGPU(xcodeml);
-    } else if(Context.isTarget(Target.CPU) || Context.isTarget(Target.ARM)) {
-      // Choose which mode to use
-      switch(Configuration.get().getParameter(Configuration.CPU_STRATEGY)) {
-        case Configuration.CPU_STRATEGY_FUSION:
-          transformFusionForCPU(xcodeml);
-          break;
-        case Configuration.CPU_STRATEGY_SINGLE:
-        default:
-          transformForCPU(xcodeml);
-          break;
-      }
-    } else {
-      throw new IllegalTransformationException("Unsupported target " +
-          Context.get().getTarget(), _claw.getPragma().lineNo());
-    }
+  /**
+   * This method should be call by any class inheriting this class to apply
+   * the last steps fo the transformation common to all SCA transformation.
+   *
+   * @param xcodeml Current translation unit.
+   */
+  void finalizeTransformation(XcodeProgram xcodeml)
+      throws IllegalTransformationException
 
+  {
     if(!_fctType.getBooleanAttribute(Xattr.IS_PRIVATE)) {
       FmoduleDefinition modDef = _fctDef.findParentModule();
       if(modDef != null) {
@@ -387,672 +374,6 @@ public class Parallelize extends ClawTransformation {
             false);
       }
     }
-  }
-
-  /**
-   * Apply GPU based transformation.
-   *
-   * @param xcodeml Current XcodeML program unit.
-   */
-  private void transformForGPU(XcodeProgram xcodeml)
-      throws IllegalTransformationException
-  {
-    AcceleratorConfiguration accConfig = Configuration.get().accelerator();
-
-    // TODO nodep should be passed in another way.
-    int collapse = Directive.generateLoopSeq(xcodeml, _fctDef,
-        CompilerDirective.CLAW.getPrefix() + " nodep");
-
-    if(!Body.isEmpty(_fctDef.body())) {
-      /* Create a nested loop with the new defined dimensions and wrap it around
-       * the whole subroutine's body. This is for the moment a really naive
-       * transformation idea but it is our start point.
-       * Use the first over clause to create it. */
-      NestedDoStatement loops =
-          new NestedDoStatement(_claw.getDimensionValuesReversed(), xcodeml);
-
-      /* Subroutine/function can have a contains section with inner subroutines
-       * or functions. The newly created (nested) do statements should stop
-       * before this contains section if it exists. */
-      Xnode contains = _fctDef.body().matchSeq(Xcode.F_CONTAINS_STATEMENT);
-      if(contains != null) {
-
-        Xnode parallelRegionStart =
-            Directive.findParallelRegionStart(_fctDef, null);
-        Xnode parallelRegionEnd =
-            Directive.findParallelRegionEnd(_fctDef, contains);
-
-        Body.shiftIn(parallelRegionStart, parallelRegionEnd,
-            loops.getInnerStatement().body(), true);
-
-        contains.insertBefore(loops.getOuterStatement());
-      } else {
-        // No contains section, all the body is copied to the do statements.
-        Xnode parallelRegionStart =
-            Directive.findParallelRegionStart(_fctDef, null);
-        Xnode parallelRegionEnd =
-            Directive.findParallelRegionEnd(_fctDef, null);
-
-        // Define a hook from where we can insert the new do statement
-        Xnode hook = parallelRegionEnd != null
-            ? parallelRegionEnd.nextSibling() : null;
-        Body.shiftIn(parallelRegionStart, parallelRegionEnd,
-            loops.getInnerStatement().body(), true);
-
-        // Hook is null then we append the do statement to the current fct body
-        if(hook == null) {
-          _fctDef.body().append(loops.getOuterStatement());
-        } else {
-          // Insert new do statement before the hook element
-          hook.insertBefore(loops.getOuterStatement());
-        }
-      }
-
-      // Prepare variables list for present/pcreate clauses and handle
-      // promotion/privatize local strategy
-      List<String> presentList =
-          Directive.getPresentVariables(xcodeml, _fctDef);
-      List<String> privateList = Collections.emptyList();
-      List<String> createList = Collections.emptyList();
-      if(accConfig.getLocalStrategy() == AcceleratorLocalStrategy.PRIVATE) {
-        privateList = Directive.getLocalArrays(xcodeml, _fctDef);
-        // Iterate over a copy to be able to remove items
-        for(String identifier : new ArrayList<>(privateList)) {
-          if(_promotions.containsKey(identifier)) {
-            privateList.remove(identifier);
-          }
-        }
-      } else if(accConfig.getLocalStrategy()
-          == AcceleratorLocalStrategy.PROMOTE)
-      {
-        createList = Directive.getLocalArrays(xcodeml, _fctDef);
-        for(String arrayIdentifier : createList) {
-          _arrayFieldsInOut.add(arrayIdentifier);
-          PromotionInfo promotionInfo = new PromotionInfo(arrayIdentifier,
-              _claw.getDimensionsForData(arrayIdentifier));
-          Field.promote(promotionInfo, _fctDef, xcodeml);
-          _promotions.put(arrayIdentifier, promotionInfo);
-
-          Field.adaptArrayRef(promotionInfo, _fctDef.body(), xcodeml);
-          Field.adaptAllocate(promotionInfo, _fctDef.body(), xcodeml);
-        }
-      }
-
-      // Generate the data region
-      Directive.generateDataRegionClause(xcodeml, presentList,
-          createList, loops.getOuterStatement(), loops.getOuterStatement());
-
-      // Generate the parallel region
-      Directive.generateParallelLoopClause(xcodeml, privateList,
-          loops.getOuterStatement(), loops.getOuterStatement(),
-          loops.size() + collapse);
-    }
-
-    Directive.generateRoutineDirectives(xcodeml, _fctDef);
-  }
-
-  /**
-   * Apply a CPU transformation which group adjacent statements together and
-   * promote the minimum number of variables possible.
-   *
-   * @param xcodeml Current XcodeML program unit.
-   * @throws IllegalTransformationException If promotion of arrays fails.
-   */
-  private void transformFusionForCPU(XcodeProgram xcodeml)
-      throws IllegalTransformationException
-  {
-    // Apply transformation only when there is a body to modify
-    if(Body.isEmpty(_fctDef.body())) {
-      return;
-    }
-    // Variables on each "depth" of the method
-    List<Set<String>> depthVars = new ArrayList<>();
-    depthVars.add(new HashSet<String>());
-    /* A growing list of affecting variables. Starting from the ones
-     * declared by SCA and successively any variables affected by the SCA
-     * declaration */
-    Set<String> affectingVars = new HashSet<>(_promotions.keySet());
-
-    /* Extract all the variables used, this doesn't include vector's
-     * iterator variables */
-    extractInfoSmartFusion(_fctDef.body(), new AtomicInteger(), depthVars,
-        affectingVars);
-
-    // Find the block we need to transform
-    List<Set<String>> targetDepthIntersections = new ArrayList<>();
-    for(int i = 0; i < depthVars.size(); i++) {
-      Set<String> depthVar = depthVars.get(i);
-
-      if(depthVar.isEmpty()) {
-        targetDepthIntersections.add(null);
-        continue;
-      }
-      // Find intersection with other blocks
-      Set<String> intersection = new HashSet<>();
-      for(int j = 0; j < depthVars.size(); j++) {
-        // Skip itself
-        if(i == j) {
-          continue;
-        }
-        // Intersect the level set with the others
-        Set<String> copy = new HashSet<>(depthVars.get(i));
-        copy.retainAll(depthVars.get(j));
-        intersection.addAll(copy);
-      }
-
-      // All are on the same index
-      targetDepthIntersections.add(intersection);
-    }
-
-    // Gather all possible loop groups before their modification
-    List<List<Xnode>> transformations = new ArrayList<>();
-    // Apply transformation at the indicated depth
-    for(int targetDepth = 0; targetDepth < depthVars.size(); targetDepth++) {
-      Set<String> vars = depthVars.get(targetDepth);
-      if(vars.isEmpty()) {
-        continue;
-      }
-      Set<String> intersection = targetDepthIntersections.get(targetDepth);
-
-      // Variables still waiting for a promotion which aren't promoted yet
-      Set<String> promotions = new HashSet<>(intersection);
-
-      // Check only the LHS arrays
-      Set<String> vectorType = new HashSet<>(vars);
-      vectorType.retainAll(affectingVars);
-      // Influenced arrays must be promoted anyway
-      for(String var : vectorType) {
-        Xid fieldId = _fctDef.getSymbolTable().get(var);
-        FbasicType type = xcodeml.getTypeTable().getBasicType(fieldId);
-        if(type != null && type.isArray()) {
-          promotions.add(var);
-        }
-      }
-
-      // Promote
-      for(String promotion : promotions) {
-        if(_promotions.containsKey(promotion)) {
-          continue;
-        }
-        PromotionInfo promotionInfo = new PromotionInfo(promotion,
-            _claw.getDimensionsForData(promotion));
-        Field.promote(promotionInfo, _fctDef, xcodeml);
-        _promotions.put(promotion, promotionInfo);
-        Field.adaptArrayRef(promotionInfo, _fctDef.body(), xcodeml);
-        Field.adaptAllocate(_promotions.get(promotion), _fctDef.body(),
-            xcodeml);
-      }
-
-      // Transform indicated level by wrapping it in a DO loop
-      gatherGroups(_fctDef.body(), new AtomicInteger(), targetDepth,
-          affectingVars, transformations);
-    }
-
-    // Avoid double DO statement generation in nested groups
-    cleanupGroups(transformations);
-
-    // After gathering we apply the transformations
-    for(List<Xnode> groupedNodes : transformations) {
-      wrapGroupInDoStatement(groupedNodes, xcodeml);
-    }
-
-    // If the last statement is a CONTAINS, fallback to the previous one
-    Xnode lastNode = _fctDef.body().lastChild();
-    while(lastNode.opcode() == Xcode.F_CONTAINS_STATEMENT) {
-      lastNode = lastNode.prevSibling();
-    }
-    // Generate the parallel region
-    Directive.generateParallelClause(xcodeml,
-        _fctDef.body().firstChild(), lastNode);
-  }
-
-  /**
-   * Check that flagged group to be wrapped in DO statements are not nested in
-   * each other. If so, the nested group is removed from the flagged list.
-   *
-   * @param groups List of grouped nodes flagged to be wrapped with DO
-   *               statement.
-   */
-  private void cleanupGroups(List<List<Xnode>> groups) {
-    List<List<Xnode>> flaggedNested = new ArrayList<>();
-
-    for(List<Xnode> group : groups) {
-      if(group.isEmpty()) {
-        continue;
-      }
-      if(isNestedIn(groups, group)) {
-        flaggedNested.add(group);
-      }
-    }
-
-    for(List<Xnode> flagged : flaggedNested) {
-      groups.remove(flagged);
-    }
-  }
-
-  /**
-   * Check if the given group has any nested statements inside one of the
-   * other groups.
-   *
-   * @param groups List of group of statements.
-   * @param group  The group of statement to be checked.
-   * @return True if the group has any statement nested in another group.
-   */
-  private boolean isNestedIn(List<List<Xnode>> groups, List<Xnode> group) {
-    Xnode probe = group.get(0);
-    for(List<Xnode> nest : groups) {
-      if(group != nest) {
-        for(Xnode possibleAncestor : nest) {
-          if(probe.isNestedIn(possibleAncestor)) {
-            return true;
-          }
-        }
-      }
-    }
-    return false;
-  }
-
-  /**
-   * Analysis of a routine in order to extracts the information necessary in
-   * order to promote a minimal number of variables.
-   * Only variables affected by affectingVars are kept in depthVars.
-   *
-   * @param body          The body to analyze.
-   * @param depth         The depth relative to the function declaration.
-   * @param depthVars     Vars used on each depth.
-   * @param affectingVars Vars which are affected by SCA and consequently
-   *                      affect other variables.
-   */
-  private void extractInfoSmartFusion(Xnode body, AtomicInteger depth,
-                                      List<Set<String>> depthVars,
-                                      Set<String> affectingVars)
-  {
-    final List<Xnode> children = body.children();
-    for(Xnode node : children) {
-      // Handle an assignment
-      if(node.opcode() == Xcode.F_ASSIGN_STATEMENT) {
-        AssignStatement as = new AssignStatement(node.element());
-        Set<String> vars = XnodeUtil.findChildrenVariables(node);
-
-        // Check if it's affected by the promotion
-        Set<String> affectedVars = new HashSet<>(vars);
-        affectedVars.retainAll(affectingVars);
-
-        // If the intersection of the sets contain anything the assignment is
-        // affected by a previous promotions
-        depthVars.get(depth.get()).addAll(affectedVars);
-        if(!affectedVars.isEmpty()) {
-          affectingVars.add(as.getLhsName());
-          depthVars.get(depth.get()).add(as.getLhsName());
-        }
-      }
-      // IF statement content shouldn't increase depth counter
-      else if(node.opcode() == Xcode.F_IF_STATEMENT) {
-        Xnode nThen = node.firstChild().matchSibling(Xcode.THEN);
-        Xnode nElse = node.firstChild().matchSibling(Xcode.ELSE);
-        if(nThen != null) {
-          extractInfoSmartFusion(nThen.body(), depth, depthVars,
-              affectingVars);
-        }
-        if(nElse != null) {
-          extractInfoSmartFusion(nElse.body(), depth, depthVars,
-              affectingVars);
-        }
-      }
-      // Handle node containing a body
-      else if(node.opcode().hasBody()) {
-        if(depthVars.size() <= depth.get() + 1) {
-          depthVars.add(new HashSet<String>());
-        }
-        depth.incrementAndGet();
-        extractInfoSmartFusion(node.body(), depth, depthVars,
-            affectingVars);
-      }
-      // Keep going inside the new node
-      else {
-        extractInfoSmartFusion(node, depth, depthVars,
-            affectingVars);
-      }
-    }
-  }
-
-  /**
-   * Transform the content of the routine and add a DO loop only at the
-   * indicated depth and only around the affecting variables.
-   *
-   * @param body          The body to transform.
-   * @param currentDepth  The depth we currently are at
-   * @param targetDepth   The depth we need to reach and transform.
-   * @param affectingVars The variable which should be contained inside the loop
-   * @param groups        List of grouped nodes.
-   */
-  private void gatherGroups(Xnode body, AtomicInteger currentDepth,
-                            final int targetDepth, Set<String> affectingVars,
-                            List<List<Xnode>> groups)
-  {
-    final List<Xnode> children = body.children();
-    final List<Xnode> groupedNodes = new ArrayList<>();
-    nodeLoop:
-    for(Xnode node : children) {
-      // Handle an assignment
-      if(node.opcode() == Xcode.F_ASSIGN_STATEMENT) {
-        if(currentDepth.get() != targetDepth) continue;
-        Set<String> vars = XnodeUtil.findChildrenVariables(node);
-        // Statement need to be in the loop
-        if(Utility.hasIntersection(vars, affectingVars)) {
-          // Is the statement wasn't promoted and is not in a current loop
-          // block, we can avoid to add it to a loop because it is not needed.
-          AssignStatement as = new AssignStatement(node.element());
-          if(!_promotions.containsKey(as.getLhsName())
-              && groupedNodes.isEmpty() && as.getVarRefNames().isEmpty())
-          {
-            continue;
-          }
-          // If the assignment is a vector, but we don't access its elements
-          // we don't have to add it to a loop
-          if(_promotions.containsKey(as.getLhsName()) &&
-              as.getVarRefNames().isEmpty())
-          {
-            addGroupedNodes(groupedNodes, groups);
-            continue;
-          }
-          groupedNodes.add(node);
-        }
-        // Particular case, unused variable inside the body
-        else {
-          addGroupedNodes(groupedNodes, groups);
-        }
-      }
-      // IF statement my have to be contained inside
-      else if(node.opcode() == Xcode.F_IF_STATEMENT) {
-        // Enter the if in case it contains DO statements
-        if(currentDepth.get() != targetDepth) {
-          Xnode nThen = node.firstChild().matchSibling(Xcode.THEN);
-          Xnode nElse = node.firstChild().matchSibling(Xcode.ELSE);
-          if(nThen != null) {
-            addGroupedNodes(groupedNodes, groups);
-            gatherGroups(nThen.body(), currentDepth, targetDepth, affectingVars,
-                groups);
-          }
-          if(nElse != null) {
-            addGroupedNodes(groupedNodes, groups);
-            gatherGroups(nElse.body(), currentDepth, targetDepth, affectingVars,
-                groups);
-          }
-          continue;
-        }
-
-        // We need to wrap the whole IF based on the condition
-        Xnode condNode = node.firstChild().matchSibling(Xcode.CONDITION);
-        if(Condition.dependsOn(condNode, affectingVars)) {
-          groupedNodes.add(node);
-          continue nodeLoop;
-        }
-
-        // We need to wrap the whole IF based on the statement inside
-        List<Xnode> assignStatements = node.matchAll(Xcode.F_ASSIGN_STATEMENT);
-        xNodeLoop:
-        for(Xnode statement : assignStatements) {
-          AssignStatement as = new AssignStatement(statement.element());
-          if(affectingVars.contains(as.getLhsName())) {
-            // If the affected node is child of a DO, a dedicated loop
-            // group will be created.
-            Xnode pnode = statement.ancestor();
-            while(pnode.hashCode() != node.hashCode()) {
-              if(pnode.opcode() == Xcode.F_DO_STATEMENT) {
-                break xNodeLoop;
-              }
-              pnode = pnode.ancestor();
-            }
-            // Add the whole if and continue otherwise
-            groupedNodes.add(node);
-            continue nodeLoop;
-          }
-        }
-
-        // If the IF statement doesn't contains any dependency we gather the
-        // potential transformation and continue
-        addGroupedNodes(groupedNodes, groups);
-      } else if(node.opcode().hasBody()) { // Handle node containing a body
-        addGroupedNodes(groupedNodes, groups);
-        currentDepth.incrementAndGet();
-        gatherGroups(node.body(), currentDepth, targetDepth, affectingVars,
-            groups);
-      } else { // Keep going inside the new node
-        addGroupedNodes(groupedNodes, groups);
-        gatherGroups(node, currentDepth, targetDepth, affectingVars, groups);
-      }
-    }
-    addGroupedNodes(groupedNodes, groups);
-  }
-
-  /**
-   * If a group of nodes flagged to be is not empty, add it to the collection
-   * containing all the group of nodes found until now.
-   * The content of group of nodes is copied into a new list added to `groups`.
-   * The content of the `groupedNodes` is then cleared.
-   * Used by CPU smart fusion transformation.
-   *
-   * @param groupedNodes A list of adjacent (grouped) nodes to be wrapped in a
-   *                     DO statement.
-   * @param groups       List containing all group of nodes.
-   */
-  private void addGroupedNodes(List<Xnode> groupedNodes,
-                               List<List<Xnode>> groups)
-  {
-    if(groupedNodes.isEmpty()) {
-      return;
-    }
-    groups.add(new ArrayList<>(groupedNodes));
-    groupedNodes.clear();
-  }
-
-  /**
-   * Given a series of sequential nodes wrapped them in a DO statement.
-   * Used by CPU smart fusion transformation.
-   *
-   * @param groupedNodes List of nodes do envelope in a DO statement.
-   * @param xcodeml      Current translation unit.
-   */
-  private void wrapGroupInDoStatement(List<Xnode> groupedNodes,
-                                      XcodeProgram xcodeml)
-  {
-    if(groupedNodes.isEmpty()) {
-      return;
-    }
-
-    // Create the DO statement
-    NestedDoStatement loop =
-        new NestedDoStatement(_claw.getDimensionValuesReversed(), xcodeml);
-
-    // Insert DO statement into the AST and add statements in its body
-    groupedNodes.get(0).insertBefore(loop.getOuterStatement());
-    for(Xnode node : groupedNodes) {
-      loop.getInnerStatement().body().append(node, false);
-    }
-
-    Directive.generateLoopDirectives(xcodeml,
-        loop.getOuterStatement(), loop.getOuterStatement(),
-        Directive.NO_COLLAPSE);
-    groupedNodes.clear();
-  }
-
-  /**
-   * Apply CPU based transformations.
-   *
-   * @param xcodeml Current XcodeML program unit.
-   * @throws IllegalTransformationException If promotion of arrays fails.
-   */
-  private void transformForCPU(XcodeProgram xcodeml)
-      throws IllegalTransformationException
-  {
-    /* Create a group of nested loop with the newly defined dimension and wrap
-     * every assignment statement in the column loop or including data with it.
-     * This is for the moment a really naive transformation idea but it is our
-     * start point.
-     * Use the first over clause to do it. */
-    List<AssignStatement> assignStatements =
-        Function.gatherAssignStatements(_fctDef);
-
-    // Iterate over all assign statements to detect all the indirect promotion.
-    for(AssignStatement assign : assignStatements) {
-      Xnode lhs = assign.getLhs();
-      String lhsName = assign.getLhsName();
-      if(lhs.opcode() == Xcode.VAR || lhs.opcode() == Xcode.F_ARRAY_REF) {
-        /* If the assignment is in the column loop and is composed with some
-         * promoted variables, the field must be promoted and the var reference
-         * switch to an array reference */
-
-        PromotionInfo promotionInfo;
-        if(shouldBePromoted(assign)) {
-          // Do the promotion if needed
-          if(!_arrayFieldsInOut.contains(lhsName)) {
-            _arrayFieldsInOut.add(lhsName);
-            promotionInfo =
-                new PromotionInfo(lhsName, _claw.getDimensionsForData(lhsName));
-            Field.promote(promotionInfo, _fctDef, xcodeml);
-            _promotions.put(lhsName, promotionInfo);
-          } else {
-            promotionInfo = _promotions.get(lhsName);
-          }
-          // Adapt references
-          if(lhs.opcode() == Xcode.VAR) {
-            Field.adaptScalarRefToArrayRef(lhsName, _fctDef,
-                _claw.getDimensionValues(), xcodeml);
-          } else {
-            Field.adaptArrayRef(_promotions.get(lhsName), _fctDef.body(),
-                xcodeml);
-            Field.adaptAllocate(_promotions.get(lhsName), _fctDef.body(),
-                xcodeml);
-          }
-          promotionInfo.setRefAdapted();
-        }
-      }
-    }
-
-    // Hold hooks on which do statement will be inserted
-    Set<Xnode> hooks = new HashSet<>();
-
-    /* Check whether condition includes a promoted variables. If so, the
-     * ancestor node using the condition must be wrapped in a do statement. */
-    List<Xnode> conditions = _fctDef.body().matchAll(Xcode.CONDITION);
-    for(Xnode condition : conditions) {
-      if(Condition.dependsOn(condition, _arrayFieldsInOut)) {
-        Xnode ancestor = condition.ancestor();
-        Iterator<Xnode> iter = hooks.iterator();
-        boolean addHook = true;
-        while(iter.hasNext()) {
-          if(ancestor.isNestedIn(iter.next())) {
-            addHook = false;
-            break;
-          }
-        }
-        if(addHook) {
-          hooks.add(ancestor);
-        }
-      }
-    }
-
-    /* Iterate a second time over assign statements to flag places where to
-     * insert the do statements */
-    for(AssignStatement assign : assignStatements) {
-      Xnode lhs = assign.getLhs();
-      String lhsName = assign.getLhsName();
-      boolean wrapInDoStatement = true;
-
-      // Check if assignment is dependant of an if statement.
-      if(assign.isChildOf(Xcode.F_IF_STATEMENT)) {
-
-        // Gather all potential ancestor if statements
-        List<Xnode> ifStatements = assign.matchAllAncestor(Xcode.F_IF_STATEMENT,
-            Xcode.F_FUNCTION_DEFINITION);
-
-        Xnode hookIfStmt = null;
-        for(Xnode ifStmt : ifStatements) {
-          if(Condition.dependsOn(
-              ifStmt.matchDirectDescendant(Xcode.CONDITION),
-              assign.getVarRefNames()))
-          {
-            // Have to put the do statement around the if as the assignment
-            // is conditional as well.
-            hookIfStmt = ifStmt;
-          }
-        }
-
-        if(hookIfStmt != null) {
-          wrapInDoStatement = false;
-          boolean addIfHook = true;
-
-          // Get rid of previously flagged hook in this if body.
-          Iterator<Xnode> iter = hooks.iterator();
-          while(iter.hasNext()) {
-            Xnode crt = iter.next();
-            if(assign.isNestedIn(crt) || hookIfStmt.isNestedIn(crt)) {
-              addIfHook = false;
-            }
-            if(crt.isNestedIn(hookIfStmt)) {
-              iter.remove();
-            }
-          }
-
-          if(addIfHook) {
-            hooks.add(hookIfStmt);
-          }
-        }
-      }
-
-      Iterator<Xnode> iter = hooks.iterator();
-      while(iter.hasNext()) {
-        if(assign.isNestedIn(iter.next())) {
-          wrapInDoStatement = false;
-          break;
-        }
-      }
-
-      if(lhs.opcode() == Xcode.F_ARRAY_REF
-          && _arrayFieldsInOut.contains(lhsName) && wrapInDoStatement)
-      {
-        hooks.add(assign);
-      } else if(lhs.opcode() == Xcode.VAR || lhs.opcode() == Xcode.F_ARRAY_REF
-          && _scalarFields.contains(lhsName))
-      {
-        if(shouldBePromoted(assign) && wrapInDoStatement) {
-          hooks.add(assign);
-        }
-      }
-    }
-
-    // Generate loops around statements flagged in previous stage
-    for(Xnode hook : hooks) {
-      NestedDoStatement loops =
-          new NestedDoStatement(_claw.getDimensionValuesReversed(), xcodeml);
-      hook.insertAfter(loops.getOuterStatement());
-      loops.getInnerStatement().body().append(hook, true);
-      hook.delete();
-      Directive.generateLoopDirectives(xcodeml,
-          loops.getOuterStatement(), loops.getOuterStatement(),
-          Directive.NO_COLLAPSE);
-    }
-
-    // Generate the parallel region
-    Directive.generateParallelClause(xcodeml,
-        _fctDef.body().firstChild(), _fctDef.body().lastChild());
-  }
-
-  /**
-   * Check whether the LHS variable should be promoted.
-   *
-   * @param assignStmt Assign statement node.
-   * @return True if the LHS variable should be promoted. False otherwise.
-   */
-  private boolean shouldBePromoted(Xnode assignStmt) {
-    Xnode rhs = assignStmt.child(Xnode.RHS);
-    if(rhs == null) {
-      return false;
-    }
-    List<Xnode> vars = XnodeUtil.findAllReferences(rhs);
-    Set<String> names = XnodeUtil.getNamesFromReferences(vars);
-    return Utility.hasIntersection(names, _arrayFieldsInOut);
   }
 
   /**
